Move all CFG analyses into a submodule.
