+++ angr/analyses/__init__.py
@@ -1,5 +1,4 @@
-from .cfg_fast import CFGFast
-from .cfg_accurate import CFGAccurate
+from .cfg import CFGFast, CFGAccurate, CFG, CFGArchOptions
 from .cdg import CDG
 from .ddg import DDG
 from .vfg import VFG
@@ -13,7 +12,6 @@ from .dfg import DFG
 from .loopfinder import LoopFinder
 from .cfg import CFG
 from .congruency_check import CongruencyCheck
-from .cfg_arch_options import CFGArchOptions
 from .static_hooker import StaticHooker
 from .reassembler import Reassembler
 from .binary_optimizer import BinaryOptimizer
