+++ angr/analyses/cfg/__init__.py
@@ -1,0 +1,10 @@
+
+# analyses
+from .cfg_fast import CFGFast
+from .cfg_accurate import CFGAccurate
+from .cfg import CFG
+
+# things to make your life easier
+from .cfg_arch_options import CFGArchOptions
+from .cfg_utils import CFGUtils
+from .cfg_node import CFGNode
