+++ angr/analyses/cfg/cfg.py
@@ -1,7 +1,7 @@
 
 import sys
 
-from ..analysis import register_analysis
+from ...analysis import register_analysis
 from .cfg_fast import CFGFast
 
 class OutdatedError(Exception):
